crickettrack-processor/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
supabase==2.4.6
python-dotenv==1.0.1
httpx==0.27.0
opencv-python-headless==4.9.0.80
numpy==1.26.4
